+++ private/bufpkg/bufmodule/bufmodulecache/cas_module_cacher.go
@@ -77,7 +77,15 @@ func (c *casModuleCacher) GetModule(
 	if err != nil {
 		return nil, err
 	}
-	return bufmodule.NewModuleForManifestAndBlobSet(ctx, manifestFromCache, blobSet)
+	return bufmodule.NewModuleForManifestAndBlobSet(
+		ctx,
+		manifestFromCache,
+		blobSet,
+		bufmodule.ModuleWithModuleIdentityAndCommit(
+			modulePin,
+			modulePin.Commit(),
+		),
+	)
 }
 
 func (c *casModuleCacher) PutModule(