+++ CHANGELOG.md
@@ -4,6 +4,15 @@
 
 - No changes yet.
 
+## [v1.21.0] - 2023-06-05
+
+- Fix issue where locally-produced images did not have module information if the corresponding
+  module was stored in the new cache.
+- Remove `buf beta registry template`.
+- Remove `buf beta registry plugin {create,deprecate,list,undeprecate,version}` and replace with
+  `buf beta registry plugin {push,delete}`.
+- Update `buf beta price` with the latest pricing information.
+
 ## [v1.20.0] - 2023-05-30
 
 - Add `--emit-defaults` flag to `buf curl` to emit default values in JSON-encoded responses.
@@ -889,7 +898,8 @@ buf check breaking proto --against .git#branch=master,subdir=proto
 
 Initial beta release.
 
-[Unreleased]: https://github.com/bufbuild/buf/compare/v1.20.0...HEAD
+[Unreleased]: https://github.com/bufbuild/buf/compare/v1.21.0...HEAD
+[v1.21.0]: https://github.com/bufbuild/buf/compare/v1.20.0...v1.21.0
 [v1.20.0]: https://github.com/bufbuild/buf/compare/v1.19.0...v1.20.0
 [v1.19.0]: https://github.com/bufbuild/buf/compare/v1.18.0...v1.19.0
 [v1.18.0]: https://github.com/bufbuild/buf/compare/v1.17.0...v1.18.0